microserviceC/requirements.txt
Flask
flask-cors
Flask-API
Flask-Restful
pymongo
python-dotenv
dnspython
pydantic
confluent-kafka